restyled login/signup pages and made login one route
