+++ src/css/VARIABLES.scss
@@ -7,15 +7,5 @@
   --blue: #2E294E;
   --white: #EEEEEE;
 
-  --nav-height: 65px;
   --fontfamily: 'Cinzel Decorative', cursive;
 }
-
-/* purple: rgb(136, 56, 140)
-red: rgb(219, 43, 57)
-lightgrey: rgb(116, 113, 128)
-midgrey: rgb(90, 90, 90)
-darkgrey: rgb(35, 38, 46)
-offwhite: rgb(245, 248, 232)
-popcornyellow: rgb(255, 230, 149)
-*/
