+++ src/css/main.scss
@@ -10,7 +10,6 @@
 /* =========================== MAIN =========================== */
 
 body {
-  /* margin-top: var(--nav-height); */
   background-color: var(--grey);
   color: var(--white);
   font-family: var(--fontfamily);
