+++ src/css/pages/RegistrationPage.scss
@@ -1,3 +1,6 @@
-#an-id {
-  outline: none;
+.registration-container {
+  background-image: url(/images/stone-bg.jpg);
+  background-position: center;
+  height: 100vh;
 }
+
