+++ src/css/registration/LogIn.scss
@@ -1,7 +1,85 @@
-#an-id {
-  outline: none;
+#login-container {
+  border: 1px solid black;
+  border-radius: 10px;
+  background-color: rgba($color: #0A090C, $alpha: 0.7);
+  max-width: 600px;
+  box-shadow: 3px 3px 6px rgb(37, 37, 37);
+  text-align: center;
+  margin: auto;
+  position: relative;
+  top: 80px;
+
+  #title {
+    padding: 20px;
+    text-align: center;
+  }
+  
+  #login-form {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  
+    h3 {
+      padding: 10px;
+    }
+  }
+
+  input {
+    height: 30px;
+    width: 400px;
+    border-radius: 5px;
+  }
+  
+  form input.error {
+    border-color: red;
+  }
+
+  #submit-button {
+    height: 50px;
+    width: 150px;
+    border-radius: 5px;
+    border: 1px solid var(--grey);
+    background-color: rgba($color: white, $alpha: .2);
+    // color: var(--white);
+    color: white;
+    margin: 15px;
+    font-size: 1.2em;
+    transition: 1s;
+    cursor: pointer;
+  }
+
+  #submit-button:hover {
+    background-color: orange;
+  }
+
+  #submit-button:disabled {
+    background-color: #454545;
+  }
+
+  p {
+    font-size: 15px;
+    margin-bottom: 20px;
+    font-family: var(--fontfamily-alt);
+  }
+
+  #signup-button {
+    border: none;
+    background: none;
+    color: orange;
+    font-size: 15px;
+    font-family: var(--fontfamily-alt);
+    cursor: pointer;
+  }
+
+  #failedToRegister {
+    color: red;
+  }
+  
+  .warning {
+    padding-top: 15px;
+    color: red;
+    text-align: center;
+  }
+  
 }
 
-#title {
-  text-align: center;
-}
